Rewriting size and toString in PaginatedIterable
Some mule components will eventually reimplement toString (!!)
instead of invoking that message on a target collection.
That implementation will also invoke size() (!!) , which was causing that
when loggers were set to debug level, an UnsupportedOperationException was thrown.

This is a design issue that requires more development in order to provide a better support
for lazy iterables, instead of relaing on a hack - making such iterables extend AbstractCollection
to work after instanceof tests - but, by now, I have reimplmented size so that it works,
but produces a warn message indicating the innefficiency.